add: update module structure
